Ungroup unrelated helpers. Tests pass locally. Remove unused code
