+++ packages/trackerUserInfoLambda/src/userInfoTypes.ts
